streamlit/requirements.txt
streamlit==1.29.0
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
